Skip publishing backup events for received WAN events in HD (#2718)
The HDPutBackupOperation was not setting the disableWanReplicationEvent
flag in the MapOperation superclass which would cause it to always
publish the backup event, even when the operation was created as a
result of receiving an event from an another cluster. This can cause
the backup event queues to grow if WAN replication was enabled on the
target cluster.

The existing tests cover the scenario. Two assertion methods for
checking the WAN queue sizes were removed since they are superseded by
one checking the backup queue sizes as well.