Send event counts instead of events when removing WAN event backups (#2727)
Send event counts instead of events when removing WAN event backups

Previously, full events, carrying key and value, were sent to the backup
replicas to remove WAN events on successful replication. Since we
don't need the key or the value, we just send the event counts. Also, we
send the event counts for all structures in one operation. Previously,
the mechanism would wait for all backup events to be removed for a
single replicated event before proceeding to the next.