Fix WAN test target endpoint setup (#2764)
Introduced in https://github.com/hazelcast/hazelcast-enterprise/pull/2681,
the port inside the instance config is incremented each time a member is
started in WAN tests.
The target endpoint calculation mechanism relies on this port as well.
This means if we start two clusters and then dynamically add WAN
replication, the endpoints would be wrongly calculated and only one
target endpoint (the one with the largest port) would be used. Others
would fail because the connection would not be established. Tests still
work however, only slower, as WAN tolerates unavailability of an
endpoint.

We remove the manual port increment as it was temporary and we can rely
on auto-increment.