Fix test relying on WAN event ordering
The cache_wan_events_should_be_processed_in_order test relies on
ordering of WAN events. When the property maxConcurrentInvocations is
set to more than 2, ordering is not respected by design. Hence, this
test can be ignored.

Also, removed testing with maxConcurrentInvocations of 1 since it is
identical to testing with -1.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2733