Prevent reordering of WAN sync events with events in WAN queues (#2800)
Prevent reordering of WAN sync events with events in WAN queues

Since handling the WAN sync event and dequeueing WAN events from the
partition queues are now done by different threads, as opposed to the
old WAN design, reordering of events can now happen.
For instance, we are syncing a map with 1 entry. The WAN sync event is
created and enqueued for replication to the target cluster by a thread
processing the WAN sync. Meanwhile, the entry is removed and the thread
collecting the WAN events from the partition queues replicates this
remove event to the target cluster. After this, the WAN sync event is
replicated and the entry is resurrected on the target cluster.
We fix this by disallowing dequeueing WAN partition queues for the
duration of WAN sync. This is a bit broad constraint and we might
improve this in the future (e.g. allowing replicating events from WAN
partition queues for partitions or maps which are not being synced). But
as WAN sync is a rare and manual process, this is acceptable.

Also, processing of concurrent WAN sync requests may now block several
threads on the WAN executor as only one sync request may be processed at
any time. This also may be improved in the future but for now we ignore
this as concurrent WAN sync requests are also rare and the situation
should resolve itself correctly eventually.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2733