WAN batching loop to exit if nothing to collect
This is a performance optimisation for the WAN event batch collector.

Without this change once the WAN events belong to the partitions replicated to a given endpoint X is drained, the subsequent collection attempts for these partitions will take the configured BATCH_MAX_DELAY_MILLIS time. This adds unnecessary latency to the replication events belong to endpoints other than X and decreases the overall throughput of the WAN replication. Besides this latency the collector logic immediately exits when taking the partitions belong to the next endpoint after X, since the elapsed time since the last sent batch exceeds BATCH_MAX_DELAY_MILLIS. This decreases the throughput and increases the latency even further by replicating the remaining events from the WAN queues in a partition-by-partition fashion, losing the batching behavior.

This change makes the batch collector loop to exit once it iterated over an endpoint's partitions without collected any event.

Fixes #2370