Clear local WAN queue before the migration operation (#2827)
Clear local WAN queue before the migration operation

Previously, the WAN queue was not cleared, causing duplicate events. In
the case of partition anti-entropy, the backup replica could end up with
more events than the primary, causing a leak. Now, we clear the WAN
queue for the partition which is migrating and decrement the counters
which is fine since the migration operation carries the entire queue
anyway and the events will be enqueued again.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2731