Use WAN executor for Merkle offloaded comparison (#2859)
If we offload WanMerkleTreeNodeCompareOperation to the ASYNC executor,
the offloaded operations may consume the whole threadpool, blocking
other tasks, like invocation callbacks that the offloaded operation is
waiting for, in which case the comparison starts hanging. This situation
emerges if the source WAN cluster size equals to the ASYNC pool size,
because every WAN source member sends an MT compare operation to the
target cluster with its node hashes.

This problem is fixed by using async invocations for MT comparison and executing
the invocation callbacks in the WAN pool instead of the ASYNC pool, so MT
comparison operations are no longer blocking callbacks.