Ensure cache configuration is fully completed before returning (#2883)
Ensure cache configuration is fully completed before returning

Concurrent cache proxy creation invocations may return early, because
the cache config is visible in the configs map. However configuration
is not done until all subsequent config is done. The additional config
is now wrapped in a separate method and configs are now wrapped
in futures which are only completed one additional setup is done.