Fix split healing WAN issue
Prior to this commit, WAN had no split-brain support. With this change,
we reset all WAN publishers on the nodes in the smaller brain, before
merging them. This reset includes clearing all partition WAN queues and
resetting the WAN counters to zero. It is safe to do since it is the
responsibility of the merge policies to decide how to merge the two
brains hence to decide what to replicate over WAN.

(cherry picked from commit 30c79332648f0b99f3aa249a452f5575ed69028f)