WanCounterMigrationTest to start members in PAUSED initial state
The test originally constructed that it pauses the WAN replication after
the instances are started. This leaves a small window for replicating
entries. The two tests reported occasional failures terminate one member
right after the last member of the cluster is started. If there was a
replication sending backup removal ops to the terminated member, WAN
replication might get stuck until the operation times out, and this
happened in the failing cases.

The change starts the members in PAUSED state instead of pausing them
after their start, which ensures no WAN replication is made until the
point the cluster is stabilized.

The change adds a few logging lines to the WAN test code, so going
forward the logs will explain a bit more about in which phase the test is.

Fixes #2962