REST API: Support for updating the license (#2929)
* REST API: Support for updating the license

Implemented the REST API for updating the license
of a running member.

Upon changing the license, the license reminder
task gets rescheduled and the license MBean is
re-registered.

Added support for compatible 'feature replacements'
when updating to a newer version license. Specifically,
when updating from a V4 license to a V5 license,
V4 WEB_SESSIONS is now considered compatible with
V5 CLIENT_FORWARDING (it replaced WEB_SESSIONS).

Made it forbidden to 'update' from a newer version
license to an older version license (e.g. from V5 to V4).

Made the *RestClusterEnterpriseTest tests sequential
because the tests change the ENTERPRISE_LICENSE_KEY
system property.

Added a test to check that it is possible to 'update'
to the same license (checks that there are no problems
with nondeterministic expiryDate, too).

Made it possible to update to a license with a different
allowed number of nodes, if this number is at least as
large as the current one.