Rebuild Merkle trees during Hot Restart (#3009)
Rebuilding Merkle trees during Hot Restart before Hot Restart completes and while the cluster state is still PASSIVE and the partition table is fixed. Therefore, no races can happen with invoking operations on
partitions decided based on a snapshot view of the partition table. Rebuilding the trees is a local concern and does not involve remote operations or distributed coordination. The Merkle trees are built both on primary and backup partition replicas taking maps' configured backup counts into account.

If rebuilding the trees (partially or fully) fails for any reason, an ERROR level message is logged with mentioning the affected map and the number of partitions failed. The suggested workaround - restarting the affected node - is also logged. Later we may need an MC button and REST endpoint to perform rebuilds with manual action but without restarting a node. Rebuild failures do not fail the Hot Restart procedure.

Fixes #2960