Use map.delete for removing entries on WAN target cluster (#2925)
Use map.delete for removing entries on WAN target cluster

Previously, map.remove was used by default which caused the previous
entry to be returned even though it was not used. It also causes
problems if the target cluster has BINARY in-memory format and does not
have the class in the classloader.

Using map.delete solves these issues. The only behaviour change is that
the map listeners will no longer receive the old value.