Avoid deserialization when merging entries received via WAN (#2951)
Avoid deserialization when merging entries received via WAN

Even with hazelcast.wan.map.useDeleteWhenProcessingRemoveEvents set to
true, the key would unconditionally be deserialized. This not only
introduces a overhead in the case when using built-in merge policies but
also fails when the class is not present in the target cluster,
rendering WAN replication unusable. Fixed by using lazy deserialization.