Report WAN sync progress based on acknowledgement
The change makes WAN synchronization to report progress based on target
acknowledgements instead of reporting based on enqueueing WAN
synchronization events in the local cluster. This way the user sees
progress only when the events reached the target cluster.
Synchronization events are getting acknowledged when the WAN operation
carrying the event gets enqueued in the target cluster. This still
leaves it possible that in the target cluster the synchronization event
wont get processed. WAN operations trigger merge operations in the target
cluster that are possibly end in remote invocations. Since WAN operations
are throttled by acknowedgements, acknowledging when the merging is done
would make synchronization more time consuming hence the acknowledgement
is not changed.

As part of this change, a synchronization UUID has been introduced that
uniquely identifies a WAN synchronization (or consistency check)
request. This UUID is visible in the logs, diagnostics and is sent to MC
as well. The UUID is useful for matching related events when
troubleshooting and also helps distinguishing between consecutive or
overlapping synchronization processes. As a downside, UUIDs make
synchronization events heavier by adding two longs to all of them.