Allow WAN replication, consistency check and sync in PASSIVE state (#2978)
Allow WAN replication, consistency check and sync in PASSIVE state

Allows WAN replication, WAN consistency check and sync with target
cluster to function while source cluster is in PASSIVE state. WAN
replication should be able to progress since it's an internal
replication mechanism and not user-initiated. Map and cache mutations
are not allowed in PASSIVE state so the queues should eventually be
drained. WAN consistency check does not mutate any data so it should be
allowed as well. WAN sync should be allowed since it does not mutate
data on the source cluster and it is manually triggered.

If the source cluster is PASSIVE and the target cluster is ACTIVE, WAN
replication and WAN sync should function normally. WAN consistency check
allows even both clusters to be PASSIVE to work.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2968