Remove WanReplicationEndpoint.collectReplicationData (#3065)
Remove WanReplicationEndpoint.collectReplicationData

This method was leaking WAN publisher implementation internals,
specifically the EWRMigrationContainer parameter indicates usage of
queues. Replaced with two methods which produce and accept any kind of
WAN event container. The exact type is specified by the
WanReplicationEndpoint - the only requirement is that it is
serializable.
As a consequence, some code was refactored out of the WAN event queue
migration operation and out of AbstractWanPublisher.