Fix JVM crash when updating record with existing value (#3079)
When calling IMap.setTtl, the existing value was used to update the
record (because we only change the TTL). HDStorageImpl.updateRecordValue
is not tolerant to this and disposes the native memory block, causing
further reads to read garbage data.

One fix would be for the DefaultRecordStore.setTtl to copy parts of the
updateRecordValue method to only update metadata but that is not enough.
The setTtl method needs to also call the map store which may mutate the
value (if it's a post-processing map store). Which sets the requirement
for the fix to make the updateRecordValue tolerant to accepting the
existing record value.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/3075