Introduce basis for WAN protocol (#2926)
Introduce basis for WAN protocol

Introduces the breaking changes necessary for exchanging enough
information between clusters when connecting for WAN replication. The
changes replace the existing AuthorizationOp and create a new
WanProtocolNegotiationOperation. The operation will exchange expected
group-name and supported WAN protocol versions. The target cluster must
then match the group-name and send a response with the chosen WAN
protocol version (if any), member and cluster version and metadata. The
returned metadata is for future use where the target cluster can return
any other kind of information which is useful in WAN replication, e.g.
target cluster partition count, partition table, if a feature is enabled
or not and so on.

Also added simple test to verify that clusters with different protocol
versions do not connect or that the highest supported protocol version
is chosen.

This PR is the simplest possible version of breaking changes necessary
in 4.0. A follow-up PR will integrate the chosen protocol version into
the serialization service (so WAN events can be serialized properly) and
will possibly introduce a better way to enumerate the supported protocol
versions (e.g. by using some service loader).