[BACKPORT] Fix WAN test for cache entry with expiration (#2921) (#3101)
Fix WAN test for cache entry with expiration (#2921)

After putting the entry with expiration in the source cluster, we assert
that it is present in the target cluster. In some cases, the entry
might have already been expired and cleaned up. We change the assertion
to assert that there were replicated entries but not the cache contents.
In addition, we add a test where we replicate entries with a very long
expiration so that we test that the cache was actually populated but we
don't wait for the entries to expire.

Fixes:
https://github.com/hazelcast/hazelcast-enterprise/issues/2912

(cherry picked from commit 29af53f563221fcd121bbebe41d804f55292c8de)