Remove WanReplicationEvent wrapper for WAN events (#3111)
Remove WanReplicationEvent wrapper for WAN events

ReplicationEventObject has most of the data necessary to replicate WAN
events. The only thing missing was serviceName and the acknowledgeType.
The service name can be inferred from the ReplicationEventObject or can
be retrieved from a method which always returns the exact name.
Acknowledge type was not transferred between clusters but was only set
in the WanReplicationEvent when processing WAN events on the target
cluster to avoid passing one more method parameter.

By removing ReplicationEventObject, we now don't need to wrap WAN
events, saving up on heap space and also avoid serializing the wrapper,
which used up bytes for the factory ID, class ID and the full string
with the service name.