Don't load permission mappings from static context
Loading it from static context prevents creating the file on
the fly. Once the SecureCallableImpl is touched, we were not able
to change the "permission-mapping.properties" file and test a
different scenario. This was forcing every instance created in single
jvm to use same properties file. After the fix, we make sure  that it
is loaded when every time a SecurityContextImpl is initialized
(This means it will loaded per hazelcast member instance)

fixes https://github.com/hazelcast/hazelcast-enterprise/issues/3098
fixes https://github.com/hazelcast/hazelcast-enterprise/issues/3097