Fix flaky MapWanSyncAPITest
Since WAN sync changed to report progress on acknowledgement, verifying
sync stats after verifying map consistency with verifyMapReplicated is
racy. There is a chance that the source and target is consistent so the
verifyMapReplicated check is passed, but the stats are not yet updated.
The fix for this is to verify the stats in assertTrueEventually.

Fixes #3124