Improved logging when adding WAN replication configuration
Improved logging for WanAddConfigTest:
- prevented repeated logging of IllegalArgumentException. This exception
is expected as we are bouncing members and at any time the call to
getNode(HazelcastInstance) can fail with a IllegalArgumentException.
This litters the log and makes finding the actual errors much harder.
Now, the exception is no longer logged.
- Logged the membership and partition table and member which is
coordinating the WAN add config operations. The tables might still be
stale when the actual operations are invoked but the changes to the
membership and partition table can be seen from other logs.
- increased overall log level to DEBUG

Also, improved logging when adding WAN replication configuration in
production code. Now, we are making a single log entry on each member
which is adding or appending WAN replication configuration. This way, we
can see if any members was omitted while broadcasting the configuration.

Related: https://github.com/hazelcast/hazelcast-enterprise/issues/2777