Dynamic metrics discovery (#3197)
This change adds dynamic metric discovery along with a few related
changes.

In order to introduce dynamic metrics, the change introduces
a distinction between static and dynamic metrics. The static metrics are
the ones that are added once and never removed during the lifetime of
the instance. These are typically JVM/OS/HZ service stats. The dynamic
stats are the ones that are discovered dynamically in every collection
cycle. The dynamic metrics are mostly data structure metrics and
networking related ones. It would have been possible to get rid of the
static metrics entirely, but collecting metrics dynamically comes with
allocation while static metrics are created once, tenure and stay there
forever without impacting GC. The preliminary measurements showed that
the static metrics consume <20K heap.

Previously to this change, consistent metric naming was introduced. It
converted the legacy metricprefix[metricid].metricname to
[id=metricid,metric=metricprefix.metricname] format. While it has
advantages, it made creating gauges more difficult and error-prone,
since any tag can be added to a metric in any order, and the whole
metric name is needed for creating the gauge. Here the old format was
more useful, so it is kept for gauges. This is done by using the old
format as the key in the MetricsRegistry's static metrics map, while the
new format is added as the name in the ProbeInstances. The metrics
collection hence uses the new format.

The gauges can be created only from static metrics cached in
MetricsRegistry. It would be possible to implement a mechanism that
updates gauges created from dynamic metrics in every collection cycle,
but currently, there are no gauges created from the metrics made dynamic.
It can be added later if needed.

The static vs dynamic distinction has made obvious by changing the
necessary methods, interface and class names.

ProbeBuilder is also renamed to MetricTagger.