Add support for excluding targets for Probes (#3204)
The change introduces excludedTargets to the @Probe annotation for
excluding probes on a certain metrics target platform. These platforms
are defined in the MetricsTarget enum, listing MANAGEMENT_CENTER, JMX,
DIAGNOSTICS.

Only probes are supported for exclusion. Metrics inserted directly via
probe functions are not. These are typically JVM/OS level metrics that
we want to publish to all platforms.

The main motivation for this change is to prevent metrics sent to
Management Center. Examples are: operation.thread*, tcp.connection[*],
tcp.inputThread*, tcp.outputThread* and internal-executor*.

This change doesn't extend any production probe with the new tag. That
change needs to be done separately after reviewing the exclusion list.