Perform file existence checks on CP Persistence restart
On CP Persistence restart, we check if the persistence directories are
valid in terms of files existing. A CP Persistence directory with CP
group directories is valid only if it also contains a local member file
that contains a CP member identity.

In addition, synchronization logic on updating locally known METADATA CP
group id is eliminated because writing to disk inside a critical section
is not a good practice. With the new logic, each METADATA CP group id is
persisted on its own file and the old file is deleted when a new id file
is written.

Last, on CP Subsystem reset, CP member file is deleted before CP group
directories. By this way, if a node crashes after deleting the CP member
file but before deleting CP group directories, it will fail on restart.