Improve corrupted entry validation during CP persistence restore
Wrote log entry length before writing the entry itself.
During restore, entry length is read first. According to read length,
if there's not enough data in the stream, then remaining data is truncated
and not restored.

Because if there's not enough available data,
most probably that was a partially written entry and not ACKed to the leader by this node.

It's also possible that "length" field is corrupted
and we are reading a wrong value. In that case, we will be truncating the log file incorrectly.
But that's something we cannot detect consistently.

If any other failure happens during restore, such as serialization error,
CRC32 validation error, EOF etc, restore process fails too.