Prevent OOME in HiDensityNearCacheTest
`HiDensityNearCacheTest#createEntryBiggerThanNativeMemory` serializes
the big paylod it tries to put to the cache. In edge cases when the
JVM doesn't have a big enough consecutive heap space for storing the
`byte[]` that the serialization service tries to create the test gets
and OOME and fails. This is fixed by creating the `HeapData` directly,
preventing the test from failing caused by the allocation above.

Fixes #3011