Hot Restart: wait until all members are ACTIVE (#3341)
Hot Restart: wait until all members are ACTIVE

When all members are done with the local load during
the Hot Restart procedure, they notify the master
which in turn tells the nodes to switch to the final
cluster state. Because this is an asynchronous operation
that is not atomic across the cluster, there is a window
of time in which some cluster members may still be
in PASSIVE state after finishing the Hot Restart data load.

This commit introduces an additional stage to the
Hot Restart procedure (HotRestartIntegrationService.startup())
that waits until all members transition from PASSIVE state.

A note about the 3.12 implementation: Because we
can't introduce an Operation to retrieve the cluster
state of a member in a patch release, I worked
around this by picking an existing Operation that:
a) does not implement AllowedDuringPassiveState and
b) is a no-op when executed.
The first property ensures that the operation fails
with a "java.lang.IllegalStateException: Cluster is
in PASSIVE state!" exception if the member is in
PASSIVE state. If the operation succeeds (with no
side-effects, thanks to the second property), it
means that the member is no longer PASSIVE.

This is an ugly hack, also with the unpleasant
side-effect of spoiling the log with (potentially
a lot of) IllegalStateException stack traces.
I attempted to reduce the log spoil by using
rather defensive idle times in the wait logic.

A proper implementation (in 4.0) should use a new
Operation to retrieve the cluster member state
and not rely on IllegalStateException etc.

Also, the issue needs to be fixed in 3.12 because
https://github.com/hazelcast/hazelcast-enterprise/issues/3325
depends on it.

If, while waiting on members to transition form the ACTIVE
state, some members do not respond within a timeout
(or if the operation fails with an unexpected exception),
we log a warning and proceed.

Fixes https://github.com/hazelcast/hazelcast-enterprise/issues/2747