License update: do not map WEB_SESSIONS to CLIENT_FILTERING (#3347)
The implementation of the dynamic license key update was
incorrectly mapping the V4 WEB_SESSIONS feature to the V5
CLIENT_FORWARDING feature. WEB_SESSIONS is a legacy feature
that was dropped in V5, and CLIENT_FILTERING is a new feature
introduced in V5. The two are unrelated.

When upgrading from an older license version to a newer license,
the code now disregards features dropped or introduced between
the two license versions, so that the old and the new licenses
are not treated as incompatible. Specific to V4->V5, the presence
of WEB_SESSIONS in the V4 license and the presence of
CLIENT_FORWARDING in the V5 license does not render the two
licenses incompatible if the other features are the same in
both licenses.

Fixes https://github.com/hazelcast/hazelcast-enterprise/issues/3346