Fix race with WAN sync record updates (#3348)
WAN sync is a coordinated effort between two threads - one fetching the
records and enqueueing them on a special queue for WAN replication and
another dequeueing them and sending them to the target cluster.

The enqueueing thread will increment the number of records to be synced
as it is enqueueing them while the dequeueing thread will decrement the
record count. If that record count reaches zero, we conclude that we
have replicated the whole partition and remove the sync stats from the
sync stats map.

But, in fact, the enqueueing thread might just be stalled and have more
records for the same partition. This may cause a NPE (because the sync
stats map does not contain the sync stats anymore), reporting that more
partitions have been completed than what the member owns (because some
partitions get completed more than once) and wrong reports about numbers
of replicated records.

The fix first adds the record count for the entire partition and then
enqueues them for replication.

Fixes:
https://github.com/hazelcast/hazelcast-enterprise/issues/3302
https://github.com/hazelcast/hazelcast-enterprise/issues/3183