Refine member metrics configuration (#3259)
* Refine member metrics configuration

This change refines the member-side metrics configuration to address two
issues with the previously added config:
- It was not clear that `retention-seconds` was for MC only
- Further improving the MC and JMX configs was not supported by the
attributes

To address these, the configuration changes as follows.

Before this change:
```
<metrics enabled="false" mc-enabled="false" jmx-enabled="false">
    <collection-interval-seconds>10</collection-interval-seconds>
    <retention-seconds>30</retention-seconds>
    <metrics-for-data-structures>true</metrics-for-data-structures>
    <minimum-level>DEBUG</minimum-level>
</metrics>
```

After this change:
```
<metrics enabled="false">
    <management-center enabled="false">
        <retention-seconds>30</retention-seconds>
    </management-center>
    <jmx enabled="false"/>
    <collection-frequency-seconds>10</collection-frequency-seconds>
    <data-structure-metrics-enabled>true</data-structure-metrics-enabled>
    <level>DEBUG</level>
</metrics>
```

The programmatic and YAML configuration follows the same structure.

It is also a requirement that all metrics configurations should be
overrideable via system properties. This is also implemented in this
change by overriding the configuration from the properties during the
`Node` instance's creation. The concrete system properties that override
the configuration fields are added to the javadocs of the metrics
config classes and to the xsd too.

Besides refining the changes above, the following gaps were also closed:
- Extends javadoc and xsd documentation for the metrics config elements
- Add Spring XML configuration
- Add config comments to `hazelcast-full-example.[xml|yaml]`
- Add metrics config to `hazelcast-fullconfig.[xml|yaml]`
- Add metrics metrics config to `hazelcast-default.[xml|yaml]`

* Disable metrics collection in MapWanSyncRESTTest

MapWanSyncRESTTest creates a Hazelcast instance with a custom Node
extension that mocks the WanService internally. If metrics collection is
enabled it may happen that the mock is accessed concurrently from
multiple threads and fails the test. Since this test doesn't verify
metrics collection, we disable metrics collection.

Additionally, the runner for this test class is replaced from
HazelcastSerialJUnitRunner to EnterpriseSerialJUnitClassRunner.

Fixes #3364