Small metrics improvements (#3374)
Two improvements/fixes:
- **Ignore descriptor mutations in `extractMetrics()`**
When decompressing and extracting a metric, mutating its
`MetricDescriptor` impacts extracting the subsequent metric too if that
metric shares at least one field with the modified one. This is fixed
by copying the extracted descriptor into `lastDescriptor` before it is
passed to the `MetricConsumer` where it can be modified.
- **Use `metricString()` instead of `toString()` in `MetricsPlugin`**
`MetricDescriptor.toString()` includes `excludedTargets` too, which is
irrelevant info for the diagnostics. Therefore,
`MetricDescriptor.metricString()` without `excludedTargets` is used
instead of `toString()`.