Cleanup leftovers for WAN replication SPI (#3363)
Cleanup leftovers for WAN replication SPI

- removed Node parameter from WanReplicationConsumer. Node is private
API and implementations may implement HazelcastInstanceAware to get
something like configuration.
- changed type of AbstractWanPublisherConfig.implementation to
WanReplicationPublisher and WanConsumerConfig.implementation to
WanReplicationConsumer.
- moved republishReplicationEvent, publishAntiEntropyEvent,
pause/stop/resume, getStats and removeWanEvents methods from public API.
These methods are all related to our WAN implementation (WAN event
republication, anti-entropy, lifecycle methods, statistics and clearing
the WAN data) and the user need not implement these methods to have the
simplest functioning WAN replication. Some of these are still exposed as
REST URIs (clear, pause/stop/resume, sync...) but the REST handlers will
be no-op in the case the user provides a custom publisher implementation.
It may be possible we expose these methods as public SPI in the future
so users may try to implement it and integrate with the REST handlers
and other subsystems.
- copied most migration and replication-related code from EE
WanReplicationService to OS. This now allows the user to fully implement
all methods on WAN publisher SPI, regardless if they are using OS or EE.
- moved all migration and replication-related methods to a
MigrationAwareWanReplicationPublisher interface. Now, if the user
wants the publisher to participate in migrations and replication, they
will need to implement this interface in addition to the
WanReplicationPublisher. Because of this, WanReplicationPublisher is
even simpler to implement.
- Move WanSyncStats and ConsistencyCheckResult to private API

Leftovers:
- code duplication between OS and EE migration and replication related
code. This can be addressed in a minor release.
- exposing event data as public methods on WAN replication events