Fix for getDistributedObjects() after Hot Restart (#3343)
* Fix for getDistributedObjects() after Hot Restart

IMap proxies created during Hot Restart are not
initialized. (This is done on purpose because index
creation - a part of map initialization - is performed
on all partitions, which would fail during Hot Restart.)
As a consequence, these proxies are invisible to
`getDistributedObjects()` since it ignores uninitialized
proxies.

The IMap proxies created during Hot Restart also aren't
published to the other cluster members (this would
trigger proxy creation/initialization on the receiving
members while they may still be busy with the Hot
Restart, which might fail for the same reasons as above).

I have extended `HotRestartIntegrationService.start()`
to force-initialize any uninitialized proxies and publish
those to the other members. This is done at a safe point
after the restarter pipeline is finished on all members
and when all members transition from the PASSIVE state.

Initializing the proxies ensures that they become visible
to `getDistributedObjects()` without having to explicitly
touch them using `getMap(...)` etc. Publishing the proxies
ensures that the proxies become visible on other members
(if they do not exist there already).

A consequence of the above change is that index creation
is now part of the Hot Restart. Previously, indexes would
be created on the first access to a map after Hot Restart
(`getMap(...)`). On one hand, the additional cost of
creating indexes for all maps is now making Hot Restart
more expensive, but on the other hand, it also makes the
first application-level access to indexed maps much
cheaper, which I believe is a good thing.

We also do not create 'implicit' proxies in EnterpriseMapServiceContextImpl
and EnterpriseCacheService once the Hot Restart procedure
is completed. Implicit proxy creation is only necessary
during data load - once the members are ready, proxies
can are created the usual way (on distributed object
creation using the public API).

This change eliminates the various race conditions with
force-initialization of proxies once the members are
ready after Hot Restart (such as the AddIndexOperation
creating uninitialized proxies on members already done
with force-initializing 'their' proxies, etc.).

Fixes https://github.com/hazelcast/hazelcast-enterprise/issues/3325