Fix for getDistributedObjects() after Hot Restart (#3400)
IMap proxies created during Hot Restart are not
initialized. (This is done on purpose because index
creation - a part of map initialization - is performed
on all partitions, which would fail during Hot Restart.)
As a consequence, these proxies are invisible to
`getDistributedObjects()` since it ignores uninitialized
proxies.

The IMap proxies created during Hot Restart also aren't
published to the other cluster members (this would
trigger proxy creation/initialization on the receiving
members while they may still be busy with the Hot
Restart, which might fail for the same reasons as above).

I have extended `HotRestartIntegrationService.start()`
to force-initialize any uninitialized proxies and publish
those to the other members. This is done at a safe point
after the restarter pipeline is finished on all members
and when all remote data-members transition to the final
cluster state (other than PASSIVE).

Initializing the proxies ensures that they become visible
to `getDistributedObjects()` without having to explicitly
touch them using `getMap(...)` etc. Publishing the proxies
ensures that the proxies become visible on other members
(if they do not exist there already).

A consequence of the above change is that index creation
is now part of the Hot Restart. Previously, indexes would
be created on the first access to a map after Hot Restart
(`getMap(...)`). On one hand, the additional cost of
creating indexes for all maps is now making Hot Restart
more expensive, but on the other hand, it also makes the
first application-level access to indexed maps much
cheaper, which I believe is a good thing.

Fixes https://github.com/hazelcast/hazelcast-enterprise/issues/3325