Update WAN sync stats with empty partitions (#3417)
This commit fixes two issues in the prod and the test code.

1) In prod code, we didn't book the empty partitions properly. These
partitions should be counted too and the MC progress update events
need to be emitted too. This was one part of the test failure that this
commit fixes. It was an obvious problem with the FULL sync, but Merkle
tree based syncing also suffered from the same problem. In that case
it happens only in a rare edge case if the consistency check finds
differences on a given partition, but between the consistency check
and syncing the difference the entries to sync got removed. This
situation can not be simulated within our tests and was caught by
reviewing the MT code in light of the finding with the FULL sync. There
are ideas on how to make this testable in the PR builder (by introducing
WAN interceptor), but making such a change at this stage of 4.0 is not
desired.

2) The order of the verifyAllPartitionsSynced and the
verifyManagementCenterInteraction methods made us skipping a block that
is supposed verify that syncedPartitions on the update events increases
monotonically and without gaps. This was a problem because the latter had
a side-effect: it extracted the MC update events that were used by the
former method to perform the validation. This is solved by moving the
verification to the verifyManagementCenterInteraction method. With this
change we could remove the map used by the test for capturing the update
events.

Fixes #3286