Add WAN config on lite members and don't send WAN sync to lite members (#3600)
Add WAN config on lite members and don't send WAN sync to lite members

When adding WAN configuration, the configuration is added as partition
operations (to avoid races with map mutation). This skips lite members
and members not owning any partitions. Because of this, we have the
following issues:
- every time WAN config is added on lite members, it is reported as
successful and not as ignored as it should be
- if WAN sync for that WAN config is invoked on that lite member, it
will throw an error because the config was not found

Because the possibilities of fixing the issue are limited in a patch
release, we go with the best-effort approach:
- in addition to sending out partition operations, the WAN config is
added locally. That means it will not be added on other members that
don't own partitions and WAN sync may still fail on those members until
the config is added on those members manually or until they are promoted
to data members, at which point they will receive replication
operations which contain the dynamically added config.
- during WAN sync, we don't send out the WAN sync request from the
coordinator to lite members. They don't own any partitions anyway and as
mentioned above, they might not even have the configuration because of
which they will throw an error. The error can be ignored and WAN sync
will be successful but the error will still be misleading.

If the cluster is on 4.1, the fix is that we send out the
AddWanConfigOperation to all members after sending the partition
operations but without requiring backups.

Newly added members should get the config as well as part of existing
post-join operations or migration/replication operations.