[4.1] Introduce WAN throttling (#3632) (#3638)
* [FORWARD-PORT] Introduce WAN throttling (#3632)

Introduce WAN throttling

Currently, there is no throttling logic present in WAN, which may allow
the WAN publishers to overload the WAN consumers if transferring the WAN
events takes less time than processing them in the target cluster. This
can be the case with WAN synchronization if the network latency is low
enough.

This PR introduces throttling on the consumer side by optionally
delaying the WAN ACKs if unhealthy congestion of invocation is detected
on a given target member. If delaying is needed the health condition is
tested periodically, with an exponential backoff.

Not addressed in this change:
Even with this throttling in place, publishers can overload the
consumers, since the number of the potential in-flight WAN batches scale
linearly with the size of the source cluster. This aspect needs to be
addressed separately. Tracked by #3622.

Changes compared to the original 3.12.z version:
- No partition double-completion elimination is needed for WAN sync in
the sync support classes since the WAN sync tracking changes introduced
in 4.0 take care of that.
- The acknowledger classes take plain Operation parameters and not
WanOperation type since the latter is a marker interface in 4.0.

* Assert sync state eventually in MapWanSyncAPITest

Co-authored-by: Zoltán Baranyi <blazember@gmail.com>