Take Merkle tree node size into account when draining WAN queues
During Merkle tree based WAN synchronization we synchronize all entries
under a given leaf together as an atomic synchronization unit. This has
the consequence that MT synchronization can easily overfill a WAN batch,
since until now the batch didn't take into account the size of the
MerkleTreeNodeEntries event and counted it as a single entry in the batch.
This makes sizing the batch difficult since WAN replicating uses the
configured batch size, while MT sync can be easily multiples of that
size, depending on the average number of the entries under the leafs.
This has the further consequence that overwhelming the target cluster -
that we introduced the WAN backpressure against recently - can be done
more easily with WAN syncing than with full sync, if the difference
between the two clusters is significant.

Considered two approaches for avoiding the above problem. The first was
removing the MerkleTreeNodeEntries collection and fill the batches with
single entries by "unpacking" these collections. While this is would be
a solution to the problem, with this we would loose the atomic
replication units, which might allow implementing advanced features,
such as detecting deletions.

The chosen implementation is modifying how the WAN synchronization queue
(shared across all partitions) is drained. Now it takes the size of the
MerkleTreeNodeEntries into account. This introduces two possible conditions
worth mentioning:
1) The batches will be potentially not entirely populated, just close
enough. This happens if adding the next MerkleTreeNodeEntries from the
synchronization queue would overfill the batch. This can be improved by
scanning the entire WAN queue for MerkleTreeNodeEntries instances that
fit in the batch. This however in the ideal configurations where a leaf
contains only a few entries wouldn't justify the complexity of the
draining algorithm.
2) If a MerkleTreeNodeEntries isntance contains more entries than the
configured batch size, we send batches bigger than the configured batch
size to make progress. This however is not a newly introduced condition,
and is very unlikely since the whole point of configuring the MTs is
finding the sufficient granularity per leaf, which ideally results the
leafs to contain only a few entries.