Replace libvmem with Memkind (#3682)
The libvmem library is replaced with the Memkind native library for
supporting volatile persistent memory use cases. Besides replacing the
library the Makefile is replaced with CMake in order to use IDEs that
support CMake for development and debugging, primarily targeting CLion.

The other significant change is giving up on static linking and bundling
the dependency .so libs in our EE jar. The reason for this is that
Memkind has secondary dependencies and Memkind itself plus some of its
dependency static libs would need to be recompiled to produce position
independent code in order to include in a single .so file. For libdaxctl
there is not even a static lib available. Going the static way would be
a much more complex build process to maintain, therefore we chose bundling
the dependencies with modifying the NEEDED and RUNPATH sections of the
produced libhazelcast-pmem.so in order to load the dependencies from the
jar file. Please see the build.md for the complete list of the
dependencies and the source of their resolution. All of the dependency
.so files are currently latest versions as downloaded from Ubuntu
mirrors.

Beside still supporting the PMEM backend for HD as it was supported in
4.0, with Memkind we now support other use cases and allocation sources
as well:
- KMEM DAX mode: for allocating PMEM memory when it is configured in
system memory mode, that is when it is available for all programs as
opaque system memory. In this mode, we allocate directly from persistent
memory. This is currently available via an undocumented system property,
will be changed to proper configuration later.
- DRAM: Regular DRAM based allocations with standard page size. This
support is not planned to be productionized in 4.1. This support is
added to test the integration with Memkind on regular machines, because
allocating DRAM shares the exact same path as KMEM DAX mode. This mode
available only through an undocumented system parameter.
- DRAM huge pages: same as DRAM, but with allocations from a huge pages
pool. This mode was added for experimental purpose to test HD with huge
pages.