Fix WanAddConfigTest#testConfigAddingWithLiteMembersAdded (#3717)
WAN config is added as a partition operation. Lite members don't own any
partitions which is why they don't get the WAN config added. In 3.12.z
and 4.0.z we applied a best-effort fix where we add the config locally
in case the member adding the config is a lite member while in 4.1 we
add a new operation which adds the config on lite members specifically.
This test fails randomly when the member adding the config is a data
member and passes when the member adding the config is a lite member.

Since 4.1 adds the config to lite members, we don't need to adapt the
test as we needed to do in 3.12.z and 4.0.z.

We apply an additional fix because the WAN queue migration operation
would add dynamically added WAN config even to a promoted lite member
but wouldn't initialize it, causing the rest of the migration operation
to fail.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/3612
Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/3598