Support multiple PMEM heaps in PMEM LibMalloc (#3723)
In 4.0 we supported only a single PMEM heap backed by a single memory
mapped file. In 4.1 we support multiple PMEM heaps and unify them into a
single one to hide this fact from the memory manager and the allocator
threads. This change offers two simple allocation strategies that direct
the allocation requests to their desired PMEM heap. These allocation
strategies are:
- Round-robin: we iterate over the PMEM heaps in a round-robin fashion,
each allocation request takes the next heap. This is a global strategy,
meaning the sequence used to implement this strategy is shared by the
allocator threads.
- NUMA-aware: if the allocator thread is bounded to a single NUMA-node
and the user configured the NUMA node on the PMEM directories backing
the heap, we pin the allocations of these bounded threads to NUMA-local
PMEM heaps to avoid interconnect traffic.

Both strategies support overflowing to other heaps if the chosen heap
has no capacity to serve the allocation requests. This enables using the
entire configured PMEM capacity even if the allocations hit the heaps in
an unbalanced fashion.

To support NUMA-aware allocation strategy, libnuma is used to detect
which NUMA node the currently running thread is bounded to.

This commit also features:
- Setting MEMKIND_HOG_MEMORY environment variable conditionally to "1", if it is not set and if "HZ_SET_HOG_MEMORY" not set to "0"
- Updated libmemkind to a manually built version with 1.10.1-rc2 + the commit we asked for that removed hole punching on PMEM kind destroy.
- Truncate the file after creation that we use check if there is enough space available on the Optane device and that if the specified directory is on a persistent memory device
- Best-effort round-robin in case we need to overflow to other PMEM heaps because the taken heap cannot serve the allocation request
- Create HDPmemMapTest as an end to end test
- Add runner annotations to PersistentMemoryPlatformTest